Konflux Onboarding Prerequisites
Add a Containerfile which packages the remoting JAR into a container
image. The remoting JAR is meant to be run in a "sidecar" container on
Kubernetes, where it establishes a connection to the main Jenkins
server. This Containerfile packages the remoting .jar into a slim
container for testing purposes. The image also makes it simple to
extract the remoting.jar file by placing it in a well-known location
(`/deployments/remoting.jar`).

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>